Add files via upload
This was supposed to be uploaded back when we had a quiz in lab class but due to confusion about the file it was saved in, I don't think I uploaded the correct one. I know the deadline has passed long ago and you strictly told us to follow the deadline but I am yet to realize that the file I uploaded was the wrong one so please try to understand my situation and check this file instead of the previous one. I deeply apologize for any inconvenience I caused and thank you in advance! Have a good day!